+++ MongoDB.Entities/Core/FuzzyString.cs
@@ -8,7 +8,7 @@ namespace MongoDB.Entities;
 
 internal class FuzzyStringSerializer : SerializerBase<FuzzyString?>, IBsonDocumentSerializer
 {
-    private static readonly StringSerializer strSerializer = new();
+    private static readonly IBsonSerializer<string> strSerializer = BsonSerializer.LookupSerializer<string>();
 
     public override void Serialize(BsonSerializationContext ctx, BsonSerializationArgs args, FuzzyString? fString)
     {